[Pure] Fix failing consistency group tempest tests
Currently CG tests have to be disabled due to this failing test. This
test will never succeed because the underlying array cannot create a
condition where this type of failure would occur.
This test has never passed due to the underlying fundamental
misunderstanding of the arrays internal logic mechanisms and should
never have been in the test suite in the first place.

Change-Id: I4923cda0e7b747c141203a024c325c4d2b8cea02