Add quota utils to cinder-manage
We have many issues with quota:

- Wrong values caused by race conditions
- Duplicated quota usage entries caused by race conditions
- Code that doesn't clean quota usage

Our current code has mechanisms to automatically refresh these values,
but there could be admins that don't want to enable it all the time, and
others that will be affected by bug #1697906 and cannot enable it.

There are a couple of scripts available online, but they all have their
issues: Not taking into account config option no_snapshot_gb_quota, not
fixing backups and groups, being racy, etc.  Moreover, they are external
tools to the OpenStack community.

We will eventually fix quotas, but in the meantime we should provide a
tool to help administrators manage the situation.

This patch adds a new 'quota' category to cinder-manage and two actions
'check' and 'sync' to allow administrators to check and fix on demand
the status of quotas (for a single or all projects).

Related-Bug: #1869749
Related-Bug: #1847791
Related-Bug: #1733179
Related-Bug: #1877164
Related-Bug: #1484343
Change-Id: Ic9323f8cfd75c0fbc425ddc9c9b35959fbe7d482