Resolve SAWarning SQLAlchemy warning
Cinder code is currently using the literal_column method when it wants
to preserver current favlue of the updated_at field, but that causes a
warning in the logs:

  SAWarning: Evaluating non-mapped column expression 'updated_at' onto
  ORM instances; this is a deprecated use case.  Please make use of

This is because the "evaluate" synchronization strategy would like to
search for objects and update them based on the UPDATE criteria passed,
however the column given, literal_column('updated_at'), is not mapped to
anything. The evaluator has to make a guess that the string contained in
these expressions should be matched to a mapped attribute on the given
entity and this guess was first removed in
https://docs.sqlalchemy.org/en/latest/changelog/changelog_12.html#change-b1e620dece39006ab44c47044e9a6fee,
then added back in
https://docs.sqlalchemy.org/en/latest/changelog/changelog_12.html#change-dff3a469788c81a46440584406cb22be
with a warning.

Closes-Bug: #1814199
Change-Id: I05d68aa420db35dc6936e7cecbe7bf8b628f6649