Fix automatic quota sync for migrating volumes
When using the automatic quota refresh via `until_refresh` and `max_age`
configuration options the calculated quota usage by the refresh will not
be correct if there are volumes that are migrating, since source and
destination volumes are counted in the refresh.

Normal quota usage calculation does not work like this, it only counts
it once, and so should the automatic quota calculation.

This patch fixes this by adding a filter to the query that skips
migration destination volumes.

It also updatest the DB implementation of volume_data_get_for_project
that didn't match the signature in cinder.db.api

Closes-Bug: #1917450
Change-Id: Ifff092917abe07726367a953f5ed420626c53bb9