Quota: Fix until_refresh config changes
until_refresh configuration option changes are only picked up by cinder
if the quota usage rows didn't exist in the DB or if until_refresh was
already set.

This means that if we enable until_refresh on an existing project then
it wouldn't be picked up.

This patch fixes this so configuration changes to until_refresh are
picked up in a timely fashion.

Closes-Bug: #1697906
Change-Id: I035ad9cdf5802b4c03c8e0fd683891f904ed321a