Fix: show volume by name for non-admins
Currently we don't pop the all_tenants field for non-admins resulting
in database not able to process it and no records are displayed.
Cinderclient passes it for most APIs since it doesn't have the context
and cannot determine if the user is admin or non-admin.
On the API side, we need to filter out the all_tenants key for non-admins
since it is only valid for admin operations.
This only happens incase of general filtering API which was introduced
in microversion 3.31, prior to this all invalid filters were removed.

Closes-Bug: #1917574
Change-Id: Idf619b4dbb64318f6e2231b13b8bacce3deaa43b
(cherry picked from commit b675b66d4f798786a30dbec00504fd2080a2e27f)