Remove unused code path in attachment_delete
In our manager's attachment_delete method we have a part of the code
that is not only used anywhere in Cinder, but it is also broken, so this
patch removes it.

Removed code was meant to allow deleting all attachments for a volume,
but the condition it's checking is on `attachment_ref` being `None`,
which can never happen since the `get_by_id` method that is assigned to
that variable will raise an exception if the entity is not found.

To fix this the condition would have to check that `attachhment_id` is
`None` instead, but since we are not using that feature we should just
remove the code.

The patch also merges `_do_attachment_delete` into `attachment_delete`
since that's the only place where it is called from.

Change-Id: I6467dc8e40d46d4158b35d40e2ef9ec273cc7e8c