PowerMax Driver - Fix for legacy PowerMax OS around generations
In the previous version of PowerMax OS generations of snapVx were
used instead of unique snap ids which were just introduced.
A generation can be returned as a 0 integer which equates to False
in python.  The fix is to convert the int to a string if it is
returned from REST as an int.

Closes-Bug: #1938572
Change-Id: I5b660776190f3026296d6d3237bd3b0d609f769f