Fix invalid yaml in reno template
This fixes two issues with the reno yaml template.

The prelude paragraph has a line longer than 80 chars.

Add a newline for the bug # entry.  This prevents an error
where yamllint complains that a token cannot start with
a backtick, if the bug # entry is used as the first item
in the "fixes" section.

Change-Id: I743e9999a42c6a7a978c88b16f02b7c5e6e826c7