Quota: Fix multiple race conditions
This patch fixes 2 race condition areas and adds a comment on why we are
not fixing another 2.

Both race conditions that we fix could have been fixed with a
"with_for_update", but one of them is more effciently handled doing a
single query update instead of retrieving the data and deleting one by
one each row.

Change-Id: I2900c8b4c06ce1a4c2bd6bb5da5274883f6a01fb