Resolve SADeprecationWarning for joinedload_all
Method joinedload_all was deprecated in SQLAlchemy in version 0.9.0 and
has been removed in version 1.4.0.

Standard way of fixing this is by replacing joinedload_all with a custom
method that chains joinedload calls for each word between ".".

In Cinder's case it's not even necessary, since all our joinedload_all
should have been joinedload in the first place.

This patch replaces joinedload_all with joinedload to resolve the
SADeprecationWarning issue and failure in newer releases.

Closes-Bug: #1832164
Change-Id: I50dc67b12764e6baa0ef05983242029b1f3d765b