PowerMax Driver - Fix for create group from source
Fix for creating a group from the source when the source
storage group is replicated. The newly created target group
must not be replicated initially so the replicate group API can
succeed.  A few force flags are also required to add and
remove from a replicated storage group.  The modify snap API
also needs a force flag.

Change-Id: I46e676cfd885a38008772f8ae20742da4955ae81