Always constraint dependencies in tox
For most of our tox jobs we use the usedevelop, which means that tox
will do a pip install -e on our source code on its own.

This means that first it will install deps (which has the constraints)
and then cinder without constraints, since they are defined in deps, and
will pull cinder dependencies ignoring the constraints.

This patch changes how we install dependencies for all tox jobs except
the lower-constraints one, and will make sure that they are always
properly constrained by changing our install_command to include the -c
option.

Change-Id: Ic9a6ac412a334710eb5e45935cd301ca80a5edb9