Change snapshots type with volume retype
Both volumes and snapshots have a volume_type_id field in the DB, but
when we migrate a volume we leave the snapshots with the old type.  This
can only happen for retypes that don't do migration, since we cannot do
a retype with migration if the volume has snapshots.

Leaving the snapshots with the old time makes a mess of the quota usage
when we do the retype as well as when we delete the snapshots.

This patch fixes the quota issue by making sure the snapshots are
retyped as well.  This means that we will check quotas for the snapshots
when retyping a volume that have them and we will properly reserve and
set the quota on retype.

Closes-Bug: #1877164
Change-Id: I90e9f85d192e1f2fee4ec8615a5bc95851a90f8e
(cherry picked from commit b77fb72aa9a58668b2c39f8c9eff06f920d76a74)