db: Add all migrations to initial alembic migration
This is kept separate to allow comparison of the initial state of the
sqlalchemy-migrate and alembic migrations as well as the final state.
This was compared with the sqlalchemy-migrate migrations using the same
pattern as described in the previous change in the series, change
I3f24313b5260a728dc98d47143ce83dba221c257.

Change-Id: I4e700aa22e192df928a07a8c18e80ea5bbf25e9b
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>