PowerMax Driver - Re-use existing initiator group/host.
When we query an initiator group/host based on it's contained
initiator(s) it will not be found unless there is an entry
in the login table.  It is possible for an initiator
group/host to exist and not be in the login table, so when we
attempt to create it using it's OpenStack name
OS-<shortHostName>-<protocol>-IG, it may fail with an
'Host already exists' error, and need to be manually deleted.

This fix checks for the existence of the initiator group/host
based on it's OpenStack name OS-<shortHostName>-<protocol>-IG.
If it already exists, we will check that the contained
initiator(s) match those in the connector object and reuse it.

Change-Id: I925697cca1a4a10dcccd630dc14d24c1735b3a0a