[SVF] Update rccg details for mirror volumes
[Spectrum Virtualize Family] For a volume with replication enabled,
during volume creation, _update_replication_properties are fetching
relationship details and updating to metadata except for rccg
properties as that volume is not added to a group. Updating IBM
Storwize cinder driver to update rccg details of a volume to metadata
when it is added to a group or removed from a group.

Closes-Bug: #1926491

Change-Id: Id3f316d58c3aa7d7930ad8d1cbb96409a39e1ba9