RBD: Don't flatten temporary resources
There are instances where cinder needs to create a temporary volume and
this can trigger a flatten of the new temporary volume, which will make
the operation take a lot longer.

In some cases this means slower operations, but in others it leads to
rpc timeout failures.

A case where we see timeout failures is when doing a backup of a
snapshot and we have rbd_flatten_volume_from_snapshot=true.

This patch ensures that we don't flatten temporary volumes.

Closes-Bug: #1916843
Change-Id: I8f55c3beb2f8df5b2227506f82ddf6ee57c951ae
(cherry picked from commit e726c07948138f514706cc69440971a2105c2bc0)