Modify manner of retrieving volume_ref
The type of the input parameter connector is dict, but as I
would like to get connector from volume_ref, the type of the
latter is str which does not comply with the type of the former
. For example, I use a list comprehension which is [a.connector
 for a in volume_ref.volume _attachment] and the elements in the
list belong to the type of str. And in such scenario, if we want
to get some values by key such as 'host', the former can utilize
connect['host'] but the latter can not. That causes a confusion.

Therefore, Modify manner of retrieving volume_ref so that the type
of connector taken from volume_ref is a instance of dict.

Change-Id: If187cfc9c79f08c5ecc1ea15c6b11fc50acbaa81