PowerMax Driver - QoS should not be set on parent storage group
The parent storage group is just a hold all for the child
storage groups which contain attributes like service level and QoS.
Setting QoS on the parent storage group can cause problems as
having a higher I/O limit on the child than the parent is not
allowed and this is not the intention.  The fix is to set
QoS on the child storage groups only.

Closes-Bug: #1930290
Change-Id: Ib2c1a44086da1fa3c64611344221c9307be22296