Netapp ONTAP: Add support to revert to snapshot
This patch adds support to revert to snapshot for NFS, FC and
iSCSI drivers with FlexVol pool.

Adds a method on client_cmode to rename a NFS volume file using
the file-rename-file zapi call.

The revert steps are:
  1. Create a clone volume file/LUN from snapshot.
  2. Change the original volume file/LUN path to a temporary path.
  3. Change clone file/LUN path to the original path.
  4. Delete the volume file/LUN on temporary path.

If any step fails, the original volume file/LUN is preserved and
the clone file/LUN is deleted.

For NFS, ONTAP clone file is not supported on FlexGroup pool, in
this case the generic implementation will perform the revert to
snapshot.

Implements: blueprint ontap-revert-to-snapshot
Co-Authored-By: Fabio Oliveira <fabioaurelio1269@gmail.com>
Change-Id: I347f0ee63d8d13ff181dd41a542a006b7c10b488