PowerMax Driver - Allow for case mismatch in SGs
When checking if a storage group is a child of a parent storage
group the check is currently case sensitive. We should allow for
a pattern match that is not case sensitive. For example,
myStorageGroup should equal MYSTORAGEGROUP or mystoragegroup.

Closes-Bug: #1929429
Change-Id: I8dd114fedece8e9d8f85c1ed237c31aede907d67
(cherry picked from commit 1f65c2a92fe3931791fca81038a27f81e1ca28fd)