PowerMax Driver - Retype fix for replication
If a volume is asynchronous or metro it can belong to a
management group as well as default storage group.  This
fix makes sure the service level of the default storage
group is returned and not the management group.
Added warnings if an array is not compression capable.

Change-Id: Ie4694d52d592e596b9f8cfda575557e5c8d24533
(cherry picked from commit 6ab6f711604f04cd63075e8658c1100ddca38f2f)